Move all security role constants into one file. Remove reference to WorkflowEditor (now on workflow pacakge)
