+++ packages/components/src/index.ts
@@ -112,7 +112,6 @@ import { LoadingSpinner } from './internal/components/base/LoadingSpinner';
 import { InsufficientPermissionsAlert } from './internal/components/permissions/InsufficientPermissionsAlert';
 import { InsufficientPermissionsPage } from './internal/components/permissions/InsufficientPermissionsPage';
 import { BasePermissionsCheckPage } from './internal/components/permissions/BasePermissionsCheckPage';
-import { APPLICATION_SECURITY_ROLES, SITE_SECURITY_ROLES } from './internal/components/permissions/constants';
 import { NotFound } from './internal/components/base/NotFound';
 import { Page } from './internal/components/base/Page';
 import { LoadingPage } from './internal/components/base/LoadingPage';
@@ -320,8 +319,10 @@ import { BasePermissions } from './internal/components/administration/BasePermis
 import { GroupManagement } from './internal/components/administration/GroupManagement';
 import { showPremiumFeatures } from './internal/components/administration/utils';
 import {
+    APPLICATION_SECURITY_ROLES,
     HOSTED_APPLICATION_SECURITY_ROLES,
     SECURITY_ROLE_DESCRIPTIONS,
+    SITE_SECURITY_ROLES,
 } from './internal/components/administration/constants';
 import { searchUsingIndex } from './internal/components/search/actions';
 import { SearchResultsModel } from './internal/components/search/models';
