+++ packages/components/src/internal/components/administration/constants.ts
@@ -31,5 +31,18 @@ export const HOSTED_APPLICATION_SECURITY_ROLES = Map<string, string>([
     [PermissionRoles.Editor, 'Editor'],
     [PermissionRoles.EditorWithoutDelete, 'Editor without Delete'],
     [PermissionRoles.Reader, 'Reader'],
-    [PermissionRoles.WorkflowEditor, 'Workflow Editor'],
 ]);
+
+export const SITE_SECURITY_ROLES = Map<string, string>([
+    [PermissionRoles.ApplicationAdmin, 'Application Administrator'],
+]);
+
+export const APPLICATION_SECURITY_ROLES = Map<string, string>([
+    [PermissionRoles.ProjectAdmin, 'Project Administrator'],
+    [PermissionRoles.FolderAdmin, 'Folder Administrator'],
+    [PermissionRoles.Editor, 'Editor'],
+    [PermissionRoles.EditorWithoutDelete, 'Editor without Delete'],
+    [PermissionRoles.Reader, 'Reader'],
+]);
+
+
