--- packages/components/src/internal/components/permissions/constants.ts
@@ -1,15 +1,0 @@
-import { Map } from 'immutable';
-import { PermissionRoles } from '@labkey/api';
-
-export const SITE_SECURITY_ROLES = Map<string, string>([
-    [PermissionRoles.ApplicationAdmin, 'Application Administrator'],
-]);
-
-export const APPLICATION_SECURITY_ROLES = Map<string, string>([
-    [PermissionRoles.ProjectAdmin, 'Project Administrator'],
-    [PermissionRoles.FolderAdmin, 'Folder Administrator'],
-    [PermissionRoles.Editor, 'Editor'],
-    [PermissionRoles.EditorWithoutDelete, 'Editor without Delete'],
-    [PermissionRoles.Reader, 'Reader'],
-    [PermissionRoles.WorkflowEditor, 'Workflow Editor'],
-]);
